+++ base/acme/src/main/java/org/dogtagpki/acme/realm/ACMERealm.java
@@ -5,35 +5,10 @@
 //
 package org.dogtagpki.acme.realm;
 
-import java.security.Principal;
-import java.security.cert.X509Certificate;
+import com.netscape.cms.realm.RealmCommon;
 
 /**
  * @author Endi S. Dewata
  */
-public class ACMERealm {
-
-    protected ACMERealmConfig config;
-
-    public ACMERealmConfig getConfig() {
-        return config;
-    }
-
-    public void setConfig(ACMERealmConfig config) {
-        this.config = config;
-    }
-
-    public void init() throws Exception {
-    }
-
-    public Principal authenticate(String username, String password) throws Exception {
-        return null;
-    }
-
-    public Principal authenticate(X509Certificate[] certs) throws Exception {
-        return null;
-    }
-
-    public void close() throws Exception {
-    }
+public class ACMERealm extends RealmCommon{
 }
