+++ base/acme/src/main/java/org/dogtagpki/acme/server/ACMEEngine.java
@@ -11,7 +11,6 @@ import java.math.BigInteger;
 import java.net.URL;
 import java.security.KeyFactory;
 import java.security.MessageDigest;
-import java.security.Principal;
 import java.security.PublicKey;
 import java.security.SecureRandom;
 import java.security.Signature;
@@ -33,7 +32,6 @@ import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.ResponseBuilder;
 
-import org.apache.catalina.realm.RealmBase;
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.lang3.NotImplementedException;
 import org.apache.commons.lang3.RandomStringUtils;
@@ -53,8 +51,6 @@ import org.dogtagpki.acme.database.ACMEDatabase;
 import org.dogtagpki.acme.database.ACMEDatabaseConfig;
 import org.dogtagpki.acme.issuer.ACMEIssuer;
 import org.dogtagpki.acme.issuer.ACMEIssuerConfig;
-import org.dogtagpki.acme.realm.ACMERealm;
-import org.dogtagpki.acme.realm.ACMERealmConfig;
 import org.dogtagpki.acme.scheduler.ACMEScheduler;
 import org.dogtagpki.acme.scheduler.ACMESchedulerConfig;
 import org.dogtagpki.acme.validator.ACMEValidator;
@@ -65,6 +61,8 @@ import org.mozilla.jss.netscape.security.util.Utils;
 import org.mozilla.jss.netscape.security.x509.X500Name;
 import org.mozilla.jss.netscape.security.x509.X509CertImpl;
 
+import com.netscape.cms.realm.RealmCommon;
+import com.netscape.cms.realm.RealmConfig;
 import com.netscape.cms.tomcat.ProxyRealm;
 
 /**
@@ -72,7 +70,7 @@ import com.netscape.cms.tomcat.ProxyRealm;
  */
 public class ACMEEngine {
 
-    public static org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(ACMEEngine.class);
+    public static final org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(ACMEEngine.class);
 
     public static ACMEEngine INSTANCE;
 
@@ -81,7 +79,6 @@ public class ACMEEngine {
     private ACMEEngineConfig config;
     private ACMEPolicy policy;
 
-    private Properties monitorsConfig;
     private ACMEEngineConfigSource engineConfigSource = null;
 
     public Random random;
@@ -97,11 +94,9 @@ public class ACMEEngine {
     private ACMEIssuerConfig issuerConfig;
     private ACMEIssuer issuer;
 
-    private ACMESchedulerConfig schedulerConfig;
     private ACMEScheduler scheduler;
 
-    private ACMERealmConfig realmConfig;
-    private ACMERealm realm;
+    private RealmCommon realm;
 
     private boolean noncesPersistent;
     private Map<String, ACMENonce> nonces = new ConcurrentHashMap<>();
@@ -354,7 +349,7 @@ public class ACMEEngine {
         try (FileReader reader = new FileReader(schedulerConfigFile)) {
             props.load(reader);
         }
-        schedulerConfig = ACMESchedulerConfig.fromProperties(props);
+        ACMESchedulerConfig schedulerConfig = ACMESchedulerConfig.fromProperties(props);
 
         logger.info("Initializing ACME scheduler");
 
@@ -366,7 +361,7 @@ public class ACMEEngine {
     public void initMonitors(String filename) throws Exception {
 
         File monitorsConfigFile = new File(filename);
-        monitorsConfig = new Properties();
+        Properties monitorsConfig = new Properties();
 
         if (monitorsConfigFile.exists()) {
             logger.info("Loading ACME monitors config from " + filename);
@@ -421,60 +416,30 @@ public class ACMEEngine {
     public void initRealm(String filename) throws Exception {
 
         File realmConfigFile = new File(filename);
-
+        RealmConfig realmConfig = null;
         if (realmConfigFile.exists()) {
             logger.info("Loading ACME realm config from " + realmConfigFile);
             Properties props = new Properties();
             try (FileReader reader = new FileReader(realmConfigFile)) {
                 props.load(reader);
             }
-            realmConfig = ACMERealmConfig.fromProperties(props);
+            realmConfig = RealmConfig.fromProperties(props);
 
         } else {
             logger.info("Loading default ACME realm config");
-            realmConfig = new ACMERealmConfig();
+            realmConfig = new RealmConfig();
         }
 
         logger.info("Initializing ACME realm");
 
         String className = realmConfig.getClassName();
-        Class<ACMERealm> realmClass = (Class<ACMERealm>) Class.forName(className);
+        Class<RealmCommon> realmClass = (Class<RealmCommon>) Class.forName(className);
 
         realm = realmClass.getDeclaredConstructor().newInstance();
         realm.setConfig(realmConfig);
-        realm.init();
+        realm.start();
 
-        ProxyRealm.registerRealm(id, new RealmBase() {
-            @Override
-            public Principal getPrincipal(String username) {
-                return null;
-            }
-
-            @Override
-            public String getPassword(String username) {
-                return null;
-            }
-
-            @Override
-            public Principal authenticate(String username, String password) {
-                try {
-                    return realm.authenticate(username, password);
-                } catch (Exception e) {
-                    logger.warn("Unable to authenticate with username: " + e.getMessage(), e);
-                    throw new RuntimeException(e);
-                }
-            }
-
-            @Override
-            public Principal authenticate(final X509Certificate[] certs) {
-                try {
-                    return realm.authenticate(certs);
-                } catch (Exception e) {
-                    logger.warn("Unable to authenticate with certificate: " + e.getMessage(), e);
-                    throw new RuntimeException(e);
-                }
-            }
-        });
+        ProxyRealm.registerRealm(id, realm);
     }
 
     public void start() throws Exception {
@@ -543,7 +508,7 @@ public class ACMEEngine {
     public void shutdownRealm() throws Exception {
         if (realm == null) return;
 
-        realm.close();
+        realm.stop();
         realm = null;
     }
